+++ src/routes/threads/thread/comments/create.ts
@@ -15,15 +15,7 @@
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import {
-    linksCl,
-    RecaptchaSecret,
-    LINKS_DOMAIN,
-    threadCl,
-    domain,
-    usersCl,
-} from "../../../../lib/common";
-import { verifyCaptcha } from "../../../../lib/recaptcha";
+import { linksCl, LINKS_DOMAIN, threadCl, domain, usersCl } from "../../../../lib/common";
 import findImages from "../../../../lib/findimages";
 import { Static, Type } from "@sinclair/typebox";
 import { generate } from "generate-password";
@@ -37,6 +29,8 @@ import { sendNotification } from "../../../../lib/notifications/sendNotification
 import { CommentSchema, IntegerSchema, RTokenSchema } from "../../../../lib/schemas";
 import { sha256 } from "../../../../lib/sha256";
 import { Link } from "../../../../models/link";
+import { RateLimitOptions } from "@fastify/rate-limit";
+import RequireReCAPTCHA from "../../../../plugins/requireRecaptcha";
 
 export default (
     fastify: FastifyInstance,
@@ -61,15 +55,15 @@ export default (
                 body: schema,
                 params: paramsSchema,
             },
-            preHandler: [checkMuted],
+            preHandler: [RequireReCAPTCHA, checkMuted],
             config: {
-                rateLimit: {
+                rateLimit: <RateLimitOptions>{
                     keyGenerator: (req: FastifyRequest) => {
                         return req.user?.id ? `user${req.user.id}` : sha256(req.ip);
                     },
-                    max: 300,
-                    ban: 50,
-                    timeWindow: 1000 * 60 * 60,
+                    max: 10,
+                    ban: 5,
+                    timeWindow: 1000 * 60,
                 },
             },
         },
@@ -82,12 +76,7 @@ export default (
         ) => {
             const id = Number(req.params.id);
 
-            const { rtoken, quote } = req.body;
-
-            if (!(await verifyCaptcha(RecaptchaSecret, rtoken)))
-                return res
-                    .code(429)
-                    .send({ statusCode: 429, error: "Recaptcha token invalid." });
+            const { quote } = req.body;
 
             const user = req.user;
             if (!user)