+++ src/routes/users/resend.ts
@@ -15,14 +15,15 @@
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { RecaptchaSecret, verificationCl } from "../../lib/common";
-import { verifyCaptcha } from "../../lib/recaptcha";
+import { verificationCl } from "../../lib/common";
 import { Static, Type } from "@sinclair/typebox";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { mg, mgDomain, verifyMsg } from "../../lib/mailgun";
 import { EmailSchema, RTokenSchema } from "../../lib/schemas";
 import { sha256 } from "../../lib/sha256";
 import { Verification } from "../../models/verification";
+import { RateLimitOptions } from "@fastify/rate-limit";
+import RequireReCAPTCHA from "../../plugins/requireRecaptcha";
 
 export default (
     fastify: FastifyInstance,
@@ -39,7 +40,7 @@ export default (
         {
             schema: { body: schema },
             config: {
-                rateLimit: {
+                rateLimit: <RateLimitOptions>{
                     max: 2,
                     ban: 5,
                     // one day
@@ -49,17 +50,14 @@ export default (
                     ) => {
                         return sha256(req.body?.email);
                     },
+                    hook: "preHandler",
                 },
             },
+            preHandler: [RequireReCAPTCHA],
         },
         async (req: FastifyRequest<{ Body: Static<typeof schema> }>, res) => {
-            const { email, rtoken } = req.body;
+            const { email } = req.body;
             const hashedEmail = sha256(email);
-
-            if (!(await verifyCaptcha(RecaptchaSecret, rtoken)))
-                return res
-                    .code(429)
-                    .send({ statusCode: 429, error: "Recaptcha token invalid." });
 
             const verificationUserData = (await verificationCl.findOne({
                 type: "register",